Separate LocationAutocompleteInput with Google Maps implementation
